legacy: switch the implementation defined format on NV12
The main intention of this patch is avoiding of software encoding and
choose the best format which isn't required encoding. It helps to reduce
CPU consumption.

This patch requires the related changes in the following projects:
 - OMX;
 - DDK.

Known issue: on NV12 format two from four cameras output image in the
invalid format (the picture is "greened").

Test: manually use camera via apk and check CPU utilization, try to use
      a preview and video capturing.
Jira-Id: ADM-1824
Change-Id: I97956338f6f873ccacbf88ea8f2109365accae5d
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>