external: rework hotplug detection with uevent
One of the the `CameraProvider` tasks is detecting new plugged devices
(cameras). The current implementation uses the `inotify` mechanism to
detect a new file creation on the dev-filesystem. But unfortunately,
sometimes it does not work well due to the complex driver
initialization flow. The new USB camera is needed to be registered as
a device of different classes (`video4linux`, `media`, etc). As a
result, sometimes the file of `video4linux` is created but the driver
has not completed an initialization.

A reproducibility rate of the issue is:
- about 75% for USB2.0 socket;
- about 20% for USB3.0 socket.

It is proposed to use `uevent` events instead of `inotify` ones. Even
more, to prevent opening of not initialized device it is proposed not
opening the device after receiving an "add" event but wait for the
corresponding "bind" event.

Another possible solution of the issue: handle such kind of errors and
try to reopen file (with or without sleep). Usually 60-90 microseconds
is enough.

Test: manualy plug and unplug an USB camera and check that it is
      detected after each connection.
Jira-Id: ADM-2585
Mantis-Id: 0007256
Change-Id: I7ac09571d2f490da4ddce5b878db17a17b3558cc
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>