external: update to the current mainline version
To make maintenance of the external HAL easier in the future it is
proposed to update on the current mainline version. The most important
feature which is introduced in the new version is the support of the
Z16 format.
Code is updated only for the device folder. Providers are not updated
because they are going to be rewritten.

The following features were backported:
 - a478cef external: add support of the YUYV format;
 - 1017e02 external: change buffer size limit for MJPG.
The following features were not backported:
 - support of `V4L2_FRMSIZE_TYPE_CONTINUOUS` frame size, because this
   feature is necessary for GMSL cameras only (this feature is contained
   in the initial commit).

Test: try to use the external HAL with UVC cameras.
Jira-Id: ADM-3415
Change-Id: Iff1866de3b8dfd05a64a339da0ad5621a75d5e41
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>