external: fix permission issue for uevent socket
When uevent socket is being created the size of the buffer matters.
At first, buffer size tries to be applied via `setsockopt` with
`SO_RCVBUF` flag, but if the requested size is greater then kernel
limits it will be reapplied with `SO_RCVBUFFORCE` flag. The second
flag requires `CAP_NET_ADMIN` capability. But this capability cannot
be applied for the Camera HAL due to SELinux neverallow rule.

You can find more details in:
    ${ANDROID_BUILD_TOP}/system/core/libcutils/uevent.cpp .
The current buffer limit can be checked via:
    $ cat /proc/sys/net/core/rmem_max

Test: $ USE_CAMERA_V4L2_KINGFISHER_HAL=true make
      flash DUT, connect a USB-camera and check that is has been
      successfully detected, logcat should not contain the following
      lines:
      ```
      CamPvdrListener@2.4-external.kingfisher: Can not open a socket:\
                                               Operation not permitted.
      ```
Jira-Id: ADM-3415
Change-Id: I1865fdbc83bd250f3ce8e22f7fd99b6e1a4ec971
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>