manifest: fix missed legacy HAL and other issues
There are two problems with the manifest fragments.

The current problem is the following: the manifest for the legacy HAL
is pointed in the wrong section (for the wrong binary `service_64`
instead of `service`). As a result the legacy HAL is not started in the
runtime.

Another potential problem is raised when the first one is fixed: the
manifests in the current state cannot be merged. This is because both of
them have the same major version. To fix the second issue it is proposed
to rewrite the manifest using <fqname> items.

The `service_64` binary and it's rc-file are removed due to cleanup
reason.

Note: you can find more info about major version validation here:
      - https://android-review.googlesource.com\
            /c/platform/system/libvintf/+/401495 .

Test: 1. Select in the device manifest required HAL(s).
      2. Build an image with camera support:
      $ USE_CAMERA_V4L2_KINGFISHER_HAL=true make
      3. Flash a DUT and ensure that desired HAL(s) are available:
      # lshal | grep -i cam
Jira-Id: ADM-3415
Change-Id: I39affbe61b55ab58a5fd469fcf47805897eebb46
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>